last try

+++ db/seeds.rb
@@ -6,12 +6,17 @@
 #   movies = Movie.create([{ name: 'Star Wars' }, { name: 'Lord of the Rings' }])
 #   Character.create(name: 'Luke', movie: movies.first)
 require 'open-uri'
-require 'json'
 
-Ingredient.destroy_all
+puts "Destroy ingredients"
+Ingredient.destroy_all if Rails.env.development?
 
-ingredient = open('https://www.thecocktaildb.com/api/json/v1/1/list.php?i=list').read
-@item = JSON.parse(ingredient)
-@item['drinks'].each do |ingredient|
-  Ingredient.create(name: ingredient['strIngredient1'])
+puts "Destroy Cocktails"
+Cocktail.destroy_all if Rails.env.development?
+
+puts "Create ingredients"
+url = "https://www.thecocktaildb.com/api/json/v1/1/list.php?i=list"
+ingredients = JSON.parse(open(url).read)
+ingredients["drinks"].each do |ingredient|
+  i = Ingredient.create(name: ingredient["strIngredient1"])
+  puts "create #{i.name}"
 end
